% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{data}
\name{apport_methods}
\alias{apport_methods}
\title{List of all apportion methods}
\format{
An object of class \code{list} of length 21.
}
\usage{
apport_methods
}
\description{
Names can be used in [proporz()] or [biproportional()], the list entries
denote the actual implementation
}
\keyword{datasets}
