% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{biproporz_examples}
\alias{biproporz_examples}
\title{Prepared and cleaned vote matrices and district seats data}
\format{
An object of class \code{list} of length 4.
}
\usage{
biproporz_examples
}
\description{
Prepared and cleaned vote matrices and district seats data
}
\keyword{datasets}
