% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biproportional-lower.R
\name{divide_votes_matrix}
\alias{divide_votes_matrix}
\title{Calculate raw seat matrix}
\usage{
divide_votes_matrix(M, col_divisors, row_divisors)
}
\arguments{
\item{M}{Matrix}

\item{col_divisors}{Divisors to apply to columns}

\item{row_divisors}{Divisors to apply to rows}
}
\value{
A matrix with the same dimension as \code{M} containing non-rounded seat values
}
\description{
Apply row and column divisors to matrix to get non-rounded seat values.
Resulting \code{NA}, \code{NaN} or \code{Inf} values are set to zero.
}
\note{
This is an internal function for \code{\link[=lower_apportionment]{lower_apportionment()}}.
}
\keyword{internal}
