% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisor.R
\name{divisor_floor}
\alias{divisor_floor}
\title{Divisor method rounding down}
\usage{
divisor_floor(votes, n_seats, quorum = 0)
}
\arguments{
\item{votes}{numeric vector with number of votes for each party}

\item{n_seats}{total number of seats}

\item{quorum}{Vote threshold a party must reach. Used as quota of total
votes within a district if less than 1 otherwise as number
of votes.}
}
\description{
Also known as D'Hondt, Jefferson, Hagenbach-Bischoff method
}
\seealso{
\code{\link{proporz}}
}
