% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisor.R
\name{divisor_geometric}
\alias{divisor_geometric}
\title{Divisor method geometric rounding}
\usage{
divisor_geometric(votes, n_seats, quorum = 0)
}
\arguments{
\item{votes}{numeric vector with number of votes for each party}

\item{n_seats}{total number of seats}

\item{quorum}{Vote threshold a party must reach. Used as quota of total
votes within a district if less than 1 otherwise as number
of votes.}
}
\description{
Also known as: Hill-Huntington method
}
\seealso{
\code{\link{proporz}}
}
