% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisor.R
\name{divisor_round}
\alias{divisor_round}
\title{Divisor method standard rounding}
\usage{
divisor_round(votes, n_seats, quorum = 0)
}
\arguments{
\item{votes}{numeric vector with number of votes for each party}

\item{n_seats}{total number of seats}

\item{quorum}{Vote threshold a party must reach. Used as quota of total
votes within a district if less than 1 otherwise as number
of votes.}
}
\value{
The number of seats per party as a vector
}
\description{
Also known as: Sainte-Laguë, Webster method
}
\seealso{
\code{\link{proporz}}
}
