% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{divisors}
\alias{divisors}
\title{Get district and party divisors}
\usage{
divisors(biproportional_result)
}
\arguments{
\item{biproportional_result}{a matrix created by \code{\link{biproportional}}
or a data.frame created by \code{\link{pukelsheim}}}
}
\description{
Show the district and party divisors used for to assign seats. This method
provides easier access to divisors stored in
attributes(...)$divisors
}
\examples{
votes_matrix = matrix(c(51,60,63,98,100,102,45,120,144), nrow = 3)
district_seats = 4:6

seats_matrix = biproportional(votes_matrix, district_seats)
divisors(seats_matrix)
#> $districts
#> [1] 40.90625 67.06318 52.00000
#>
#> $parties
#> [1] 0.9694748 0.9921875 1.0234375

}
