% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biproportional-lower.R
\name{find_divisor}
\alias{find_divisor}
\title{Find divisor to assign seats}
\usage{
find_divisor(votes, divisor_from, divisor_to, target_seats, round_func)
}
\arguments{
\item{votes}{Votes (matrix with only one column or row, allows to use row/colnames
within \code{round_func})}

\item{divisor_from}{Lower bound for divisor search range (is decreased if necessary)}

\item{divisor_to}{Upper bound for divisor search range (is increased if necessary)}

\item{target_seats}{Number of seats to distribute (single number)}

\item{round_func}{Rounding function}
}
\value{
divisor (single number)
}
\description{
Find a divisor between \code{divisor_from} and \code{divisor_to} such that
\code{sum(round_func(votes/divisor))} equals \code{target_seats}
}
\note{
This is an internal function for \code{\link[=find_matrix_divisors]{find_matrix_divisors()}}.
}
\keyword{internal}
