% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biproportional.R
\name{find_lower_apport_divisors}
\alias{find_lower_apport_divisors}
\title{Find divisors for a matrix with alternate scaling}
\usage{
find_lower_apport_divisors(M, seats_cols, seats_rows, round_func)
}
\arguments{
\item{M}{votes_matrix}

\item{seats_cols}{target seats for each column}

\item{seats_rows}{target seats for each row}

\item{round_func}{rounding function, called like
\code{round_func(M/row_divisors/col_divisors)}}
}
\value{
list of divisors
}
\description{
Find divisors for a matrix with alternate scaling
}
\keyword{internal}
