% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finland2019}
\alias{finland2019}
\title{Example data from the Finnish parliamentary elections (2019)}
\format{
An object of class \code{list} of length 2.
}
\usage{
finland2019
}
\description{
Example data from the Finnish parliamentary elections (2019)
}
\keyword{datasets}
