% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grisons2022}
\alias{grisons2022}
\title{Election Data for the Cantonal Council of Grisons (2022)}
\format{
List containing:
\itemize{
\item{\code{votes_matrix} the number of votes for each district and party
(39 rows, 5 columns). Note that districts are in rows and parties
in columns, contrary to convention within this package.}
\item{\code{district_seats_df} with the number of seats per district
(37 rows, 2 columns)}
}
}
\source{
\url{https://www.gr.ch/DE/publikationen/abstimmungenwahlen/Grossratswahlen-2022/resultate/Seiten/Resultate.aspx}
}
\usage{
grisons2022
}
\description{
Example election data from the 2020 election for the cantonal council
of Grisons (Grossrat, Cussel grond, Gran Consiglio) in Switzerland. The data
has been extracted from the document "Protokoll der Grossratswahlen vom 15.
Mai 2022".
}
\keyword{data}
