% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisor.R
\name{hzv}
\alias{hzv}
\title{Highest averages method}
\usage{
hzv(votes, n_seats, divisors)
}
\arguments{
\item{votes}{vector with votes for each party}

\item{n_seats}{total number of seats}

\item{divisors}{sequence of divisors (length equal to the number of seats).
If it's a single number (e.g. 0.5), a sequence is generated
starting with it.}
}
\value{
The number of seats per party as a vector
}
\description{
Alias for \code{\link{highest_averages_method}}. Named after its
German equivalent "Höchstzahlverfahren".
}
