% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_to_df}
\alias{pivot_to_df}
\title{Pivot wide matrix to long data.frame}
\usage{
pivot_to_df(matrix_wide, value_colname = "values")
}
\arguments{
\item{matrix_wide}{matrix in wide format}

\item{value_colname}{name for the data.frame new column}
}
\value{
data.frame with 3 columns
}
\description{
Pivot wide matrix to long data.frame
}
\examples{
mtrx = matrix(1:6, nrow = 2)
pivot_to_df(mtrx)

dimnames(mtrx) <- list(party = c("A", "B"), region = c("I", "II", "III"))
pivot_to_df(mtrx, "seats")

}
\seealso{
\code{\link[=pivot_to_matrix]{pivot_to_matrix()}}
}
