% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bazi.R
\name{read_bazi}
\alias{read_bazi}
\title{Read a .bazi file}
\usage{
read_bazi(file.bazi)
}
\arguments{
\item{file.bazi}{bazi file}
}
\value{
list of bazi tokens with a parsed \code{data} data.frame and \code{seat} data.frame or number
}
\description{
Read a .bazi file
}
\note{
This function is experimental and not exported.
}
\keyword{internal}
