% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{run_app}
\alias{run_app}
\title{Use biproportional apportionment interactively in a shiny app}
\usage{
run_app(votes_matrix = NULL, district_seats = NULL)
}
\arguments{
\item{votes_matrix}{Optional \code{votes_matrix} to load upon start}

\item{district_seats}{Optional \code{district_seats} to load upon start}
}
\value{
Calling the function starts the shiny app
}
\description{
Use biproportional apportionment interactively in a shiny app
}
\examples{
if(interactive()) {
    # You need to have the packages 'shiny' and
    # shinyMatrix' installed to run the app
    run_app()

    # It's possible to load data with the app
    run_app(uri2020$votes_matrix, uri2020$seats_vector)
}
}
