% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uri2020}
\alias{uri2020}
\title{Election Data for the Cantonal Council of Uri (2020)}
\format{
List containing:
\itemize{
\item{\code{votes_matrix} the number of votes for each party and district
(4 rows, 4 columns)}
\item{\code{seats_vector} with the number of seats per district
(length 4)}
}
}
\source{
\url{https://www.ur.ch/abstimmungen/termine/9322}
}
\usage{
uri2020
}
\description{
Example election data from the 2020 election for the cantonal council
of Uri (Landrat) in Switzerland. The data has been extracted from the report
"Landratswahlen 2020: Statistische Auswertung".
}
\examples{
uri2020$votes_matrix

uri2020$seats_vector

}
\keyword{data}
