% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biproportional.R
\name{weigh_list_votes}
\alias{weigh_list_votes}
\title{Create weighted votes matrix}
\usage{
weigh_list_votes(votes_matrix, seats_district)
}
\arguments{
\item{votes_matrix}{votes matrix}

\item{seats_district}{seats per district (vector)}
}
\value{
the weighted \code{votes_matrix}
}
\description{
Weigh list votes by dividing the votes matrix entries by the number
of seats per district. No input checks are performed.
}
\examples{
vm = matrix(c(100,50,20,10), 2)
weigh_list_votes(vm, c(10, 2))
}
\keyword{internal}
