% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biproportional-upper.R
\name{weight_list_votes}
\alias{weight_list_votes}
\title{Create weighted votes matrix}
\usage{
weight_list_votes(votes_matrix, district_seats)
}
\arguments{
\item{votes_matrix}{votes matrix}

\item{district_seats}{seats per district, vector with same length
as \code{ncol(votes_matrix)}}
}
\value{
the weighted \code{votes_matrix}
}
\description{
Weight list votes by dividing the votes matrix entries by the number
of seats per district. This method is used in \code{\link[=upper_apportionment]{upper_apportionment()}} if
\code{use_list_votes} is \code{TRUE} (default).
}
\note{
The weighted votes are not rounded. Matrix and vector names are ignored.
}
\examples{
weight_list_votes(uri2020$votes_matrix, uri2020$seats_vector)

}
