% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zug2018}
\alias{zug2018}
\title{Election Data for the Cantonal Council of Zug (2018)}
\format{
An object of class \code{data.frame} with 267 rows and 49 columns.
}
\source{
Kanton Zug (01.07.2022, 10:27:58). Kantonsratswahl 2018 (CSV).
\url{https://wab.zug.ch/elections/kantonsratswahl-2018/data-csv}
}
\usage{
zug2018
}
\description{
Example election data from the 2018 election for the cantonal council
of Zug (Kantonsrat) in Switzerland.
}
\examples{
str(zug2018)

}
\keyword{data}
