% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{data}
\name{apport_methods}
\alias{apport_methods}
\title{List of all apportion methods}
\format{
An object of class \code{list} of length 21.
}
\usage{
apport_methods
}
\description{
Names can be used in [proporz()] or [biproportional()], the list entries
denote the actual implementation.
}
\details{
\code{list(
"divisor_floor" = "floor",
"d'hondt" = "floor",
"jefferson" = "floor",
"divisor_round" = "round",
"sainte-lague" = "round",
"webster" = "round",
"divisor_ceiling" = "ceiling",
"adams" = "ceiling",
"dean" = "harmonic",
"huntington-hill" = "geometric",
"hill-huntington" = "geometric",
"hare-niemeyer" = "quota_largest_remainder",
"hamilton" = "quota_largest_remainder",
"vinton" = "quota_largest_remainder",
"hagenbach-bischoff" = "floor",
"ceiling" = "ceiling",
"round" = "round",
"floor" = "floor",
"harmonic" = "harmonic",
"geometric" = "geometric",
"quota_largest_remainder" = "quota_largest_remainder"
)}
}
\keyword{datasets}
