% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{ceil_at}
\alias{ceil_at}
\title{Round x up if x-floor(x) >= threshold}
\usage{
ceil_at(x, threshold)
}
\arguments{
\item{x}{numeric value}

\item{threshold}{threshold in 0..1}
}
\description{
Round x up if x-floor(x) >= threshold
}
