% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biproportional.R
\name{divide_votes_matrix}
\alias{divide_votes_matrix}
\title{Calculate raw seat matrix}
\usage{
divide_votes_matrix(M, col_divisors, row_divisors)
}
\arguments{
\item{M}{matrix}

\item{col_divisors}{divisors to apply to columns}

\item{row_divisors}{divisors to apply to rows}
}
\value{
matrix with the same dimension as \code{M} containing non-rounded seat values
}
\description{
Apply row and column divisors to matrix to get non-rounded seat values.
}
\keyword{internal}
