% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biproportional.R
\name{find_matrix_divisors}
\alias{find_matrix_divisors}
\title{Find divisors for a matrix with alternate scaling}
\usage{
find_matrix_divisors(M, seats_cols, seats_rows, round_func)
}
\arguments{
\item{M}{votes_matrix}

\item{seats_cols}{target seats for each column}

\item{seats_rows}{target seats for each row}

\item{round_func}{rounding function. Called like
\code{round_func(M/row_divisors/col_divisors)}, divisors are applied row/col-wise with
\code{\link[=divide_votes_matrix]{divide_votes_matrix()}}.}
}
\value{
list of divisors (column and row)
}
\description{
Find divisors for a matrix with alternate scaling
}
\keyword{internal}
