% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biproportional-divisors.R
\name{get_divisors}
\alias{get_divisors}
\title{Get district and party divisors from biproporz result}
\usage{
get_divisors(biproporz_result)
}
\arguments{
\item{biproporz_result}{a matrix created by \code{\link[=biproporz]{biproporz()}} or a
data.frame created by \code{\link[=pukelsheim]{pukelsheim()}}}
}
\value{
The district and party divisors in a list, each as a vector
}
\description{
Show the district and party divisors used to assign seats.
This method provides easier access to divisors stored in
\code{attributes(...)$divisors}.
}
\examples{
seats_matrix = biproporz(uri2020$votes_matrix, uri2020$seats_vector)
get_divisors(seats_matrix)

}
