% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{get_divisors}
\alias{get_divisors}
\title{Get district and party divisors from \code{\link[=biproporz]{biproporz()}} result}
\usage{
get_divisors(biproporz_result)
}
\arguments{
\item{biproporz_result}{a matrix created by \code{\link{biproporz}}
or a data.frame created by \code{\link{pukelsheim}}}
}
\value{
The district and party divisors in a list
}
\description{
Show the district and party divisors used for to assign seats.
This method provides easier access to divisors stored in
\code{attributes(...)$divisors}
}
\examples{
votes_matrix = matrix(c(51,60,63,98,100,102,45,120,144), nrow = 3)
district_seats = 4:6

seats_matrix = biproporz(votes_matrix, district_seats)
get_divisors(seats_matrix)

}
