% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisor.R
\name{hzv}
\alias{hzv}
\title{Highest averages method}
\usage{
hzv(party_votes, n_seats, divisors)
}
\arguments{
\item{party_votes}{vector with votes for each party}

\item{n_seats}{total number of seats}

\item{divisors}{sequence of divisors. If it's a single number, a sequence
is generated starting with it.}
}
\description{
Alias for \code{\link{highest_averages_method}}. Named after its
German equivalent "Höchstzahlverfahren".
}
