% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biproportional.R
\name{lower_apportionment}
\alias{lower_apportionment}
\title{Calculate lower apportionment}
\usage{
lower_apportionment(votes_matrix, seats_cols, seats_rows, method = "round")
}
\arguments{
\item{votes_matrix}{votes matrix}

\item{seats_cols}{number of seats per column (districts/regions), calculated
from upper_apportionment()}

\item{seats_rows}{number of seats per row (parties/lists), calculated
from upper_apportionment()}

\item{method}{Apportion method that defines how seats are assigned,
see \link{proporz}. Note  that the iterative process is only
guaranteed to terminate with "round" (Sainte-Laguë/Webster
method). It is also possible to provide a function that works
like base::round(x) (i.e. can handle a matrix).}
}
\value{
seat matrix with column and row divisors stored in attributes
}
\description{
Iterate and change column and row divisors such that
\code{colSums(round(M/col_divisors/row_divisors)) == seats_col} and
\code{rowSums(round(M/col_divisors/row_divisors)) == seats_row}
}
\details{
The result is obtained by an iterative process. Initially, for each district
a divisor is chosen using the highest averages method for the votes allocated
to each regional party list in this region. For each party a party divisor is
initialized with 1.

Effectively, the objective of the iterative process is to modify the regional
divisors and party divisors so that the number of seats in each regional
party list equals the number of their votes divided by both the regional and
the party divisors.

The following two correction steps are executed until this objective is
satisfied:
\itemize{
\item modify the party divisors such that the apportionment within each
party is correct with the chosen rounding method,
\item modify the regional divisors such that the apportionment within the
region is correct with the chosen rounding method.
}
}
\note{
The iterative process in the lower apportionment is only guaranteed to
terminate with Sainte-Laguë/Webster method.
}
\references{
Oelbermann, K. F. (2016). Alternate scaling algorithm for
biproportional divisor methods. Mathematical Social Sciences, 80, 25-32.
}
\seealso{
\code{\link{biproportional}}, \code{\link{lower_apportionment}}
}
