% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_to_matrix}
\alias{pivot_to_matrix}
\title{Pivot long data.frame to wide matrix}
\usage{
pivot_to_matrix(df)
}
\arguments{
\item{df}{data.frame in long format with exactly 3 columns}
}
\value{
a matrix
}
\description{
Pivot long data.frame to wide matrix
}
\note{
This function exists because reshape is hard to handle and the package
should have no dependencies.
}
\examples{
df = data.frame(party = c("A", "A", "A", "B", "B", "B"),
                region = c("III", "II", "I", "I", "II", "III"),
                seats = c(5L, 3L, 1L, 2L, 4L, 6L))
pivot_to_matrix(df)

}
\seealso{
\code{\link[=pivot_to_df]{pivot_to_df()}}
}
