% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pivot_to_matrix}
\alias{pivot_to_matrix}
\title{Pivot long data.frame to wide matrix}
\usage{
pivot_to_matrix(df)
}
\arguments{
\item{df}{data.frame in long format with exactly 3 columns}
}
\value{
a matrix
}
\description{
Pivot long data.frame to wide matrix
}
\note{
This function exists because I wanted to have no dependencies.
Wrangling with reshape isn't fun at all and it became glaringly
apparent why better tools like tidyr exist
}
