% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proporz.R
\name{proporz}
\alias{proporz}
\title{Proportional Apportionment}
\usage{
proporz(votes, n_seats, method, quorum = 0)
}
\arguments{
\item{votes}{numeric vector with number of votes for each party}

\item{n_seats}{total number of seats}

\item{method}{Apportionment method to use. Not case sensitive. See details.}

\item{quorum}{Vote threshold a party must reach. Used as quota of total
votes within a district if less than 1 otherwise as number
of votes.}
}
\description{
Calculate seat apportionment for legislative bodies.
}
\details{
The following methods are available:
\itemize{
         \item{d'hondt, jefferson, hagenbach-bischoff, floor}
         \item{sainte-lague, webster, round}
         \item{adams, ceiling}
         \item{dean, harmonic}
         \item{huntington-hill, hill-huntington, geometric}
         \item{hare-niemeyer, hamilton, vinton, quota_largest_remainder}
}
}
\examples{
votes = c("Party A" = 651, "Party B" = 349, "Party C" = 50)

proporz(votes, 10, "sainte-lague")

proporz(votes, 10, "hill-huntington")

proporz(votes, 10, "hill-huntington", quorum = 0.05)

}
