% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proporz.R
\docType{data}
\name{proporz_methods}
\alias{proporz_methods}
\title{List of method names and their implementation}
\format{
An object of class \code{list} of length 23.
}
\usage{
proporz_methods
}
\value{
Named list of methods
}
\description{
Names can be used in \code{\link[=proporz]{proporz()}} or \code{\link[=biproporz]{biproporz()}}, the list entries
denote the name of the implementation function.
}
\keyword{internal}
