% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biproportional.R
\name{pukelsheim}
\alias{pukelsheim}
\title{Biproportional apportionment with data frames}
\usage{
pukelsheim(
  votes_df,
  district_seats_df,
  new_seats_col = "seats",
  quorum_districts = 0,
  quorum_total = 0
)
}
\arguments{
\item{votes_df}{data.frame (long format) with 3 columns: party id/name,
district id/name, and votes (actual colnames can differ)}

\item{district_seats_df}{data.frame with 2 colums: district id/name,
number of seats for a district (actual colnames
can differ)}

\item{new_seats_col}{name of the new column}

\item{quorum_districts}{Vote threshold a party must reach in _at least_ one
district. Used as quota of total votes within a
district if less than 1 otherwise as number of votes.}

\item{quorum_total}{Vote threshold a party must reach for all votes cast.
Used as quota of total votes within a district if less
than 1 otherwise as number of votes.}
}
\description{
Parties failing to reach at least one quorum have their votes set to zero.
}
\examples{
votes_df = unique(zug2018[c("list_id", "entity_id", "list_votes")])
district_seats_df = unique(zug2018[c("entity_id", "election_mandates")])

seats_df = pukelsheim(votes_df,
                      district_seats_df,
                      quorum_districts = 0.05,
                      quorum_total = 0.03)

head(seats_df)
#>   list_id entity_id list_votes seats
#> 1       2      1701       8108     2
#> 2       1      1701       2993     1
#> 3       3      1701      19389     3
#> 4       4      1701      14814     2
#> 5       5      1701       4486     1
#> 6       6      1701      15695     3

}
