% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quorum.R
\name{quorum_all}
\alias{quorum_all}
\title{Create a quorum function where \emph{all} quorums must be reached}
\usage{
quorum_all(any_district, total)
}
\arguments{
\item{any_district}{Vote threshold a party must reach in \emph{at least}
one district. Used as quota of total votes within a
district if less than 1 otherwise as number of votes.
Must be greater than 0.
Uses \code{\link[=reached_quorum_any_district]{reached_quorum_any_district()}}.}

\item{total}{Vote threshold a party must reach for all votes cast.
Used as quota of total votes if less than 1, otherwise
as number of votes. Must be greater than 0.
Uses \code{\link[=reached_quorum_total]{reached_quorum_total()}}.}
}
\value{
a function(votes_matrix) which then returns a boolean vector with
length equal to the number of lists/parties (votes_matrix rows),
denoting whether a party has reached both quorums.
}
\description{
\code{quorum_any} and \code{quorum_all} are normally used in \code{\link[=biproportional]{biproportional()}}
or \code{\link[=pukelsheim]{pukelsheim()}}. Missing quorum parameters are ignored.
}
\note{
Currently only two quorums are implemented (any_district and total).
If you only use one quorum, \code{\link[=quorum_any]{quorum_any()}} and \code{\link[=quorum_all]{quorum_all()}} are identical.
}
\examples{
votes_matrix = matrix(c(502, 55, 80, 10, 104, 55, 0, 1), ncol = 2)
dimnames(votes_matrix) <- list(c("A", "B", "C", "D"), c("Z1", "Z2"))
seats = c(Z1 = 50, Z2 = 20)

# use as parameter in biproportional (general use case)
biproporz(votes_matrix, seats)
#>   Z1 Z2
#> A 39 12
#> B  5  8
#> C  5  0
#> D  1  0

# quorum_any and quorum_all are identical if you only use one criterion
biproporz(votes_matrix, seats, quorum = quorum_any(any_district = 0.1))
#>   Z1 Z2
#> A 40 12
#> B  5  8
#> C  5  0
#> D  0  0

biproporz(votes_matrix, seats, quorum = quorum_any(total = 100))
#>   Z1 Z2
#> A 44 12
#> B  6  8
#> C  0  0
#> D  0  0

biproporz(votes_matrix, seats, quorum = quorum_any(total = 0.5))
#>   Z1 Z2
#> A 50 20
#> B  0  0
#> C  0  0
#> D  0  0
biproporz(votes_matrix, seats, quorum = quorum_any(any_district = 0.1, total = 100))
#>   Z1 Z2
#> A 40 12
#> B  5  8
#> C  5  0
#> D  0  0

biproporz(votes_matrix, seats, quorum = quorum_all(any_district = 0.1, total = 100))
#>   Z1 Z2
#> A 44 12
#> B  6  8
#> C  0  0
#> D  0  0

}
\seealso{
\code{\link{quorum_any}}
}
