% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota.R
\name{quota_largest_remainder}
\alias{quota_largest_remainder}
\title{Largest remainder method}
\usage{
quota_largest_remainder(votes, n_seats, quorum = 0)
}
\arguments{
\item{votes}{numeric vector with number of votes for each party}

\item{n_seats}{total number of seats}

\item{quorum}{Vote threshold a party must reach. Used as quota of total
votes within a district if less than 1 otherwise as number
of votes.}
}
\value{
The number of seats per party as a vector
}
\description{
Also known as:  Hamilton, Hare-Niemeyer, Vinton method
}
\examples{
votes = c(47000, 16000, 15800, 12000, 6100, 3100)
quota_largest_remainder(votes, 10)

}
\seealso{
\code{\link{proporz}}
}
