% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quota.R
\name{quota_largest_remainder}
\alias{quota_largest_remainder}
\title{Largest remainder method}
\usage{
quota_largest_remainder(votes, n_seats, throw_equal_remainder_error = TRUE)
}
\arguments{
\item{votes}{numeric vector with number of votes for each party}

\item{n_seats}{total number of seats}

\item{throw_equal_remainder_error}{Should an error be thrown if two parties
have the same remainder? Default is TRUE
since apportionment behavior is undefined.}
}
\description{
Also known as:  Hamilton, Hare-Niemeyer, Vinton method
}
\seealso{
\code{\link{proporz}}
}
