% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quorum.R
\name{reached_quorum_any_district}
\alias{reached_quorum_any_district}
\title{Check if lists/parties have reached a quorum in at least one district}
\usage{
reached_quorum_any_district(votes_matrix, quorum_districts)
}
\arguments{
\item{votes_matrix}{votes matrix}

\item{quorum_districts}{Vote threshold a party must reach in \emph{at least}
one district. Used as quota of total votes within a
district if less than 1 otherwise as number of votes.
Must be greater than 0.}
}
\value{
boolean vector with length equal to the number of lists/parties
(\code{votes_matrix} rows) whether they reached the quorum or not
}
\description{
Base implementation, used by \code{\link[=quorum_any]{quorum_any()}} and \code{\link[=quorum_all]{quorum_all()}}.
}
