% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zug2018}
\alias{zug2018}
\title{Example election data from Zug Switzerland (2018)}
\format{
An object of class \code{data.frame} with 267 rows and 49 columns.
}
\usage{
zug2018
}
\description{
Example election data from Zug Switzerland (2018)
}
\keyword{datasets}
